+++ src/lib/app/mod.rs
@@ -1,2 +1,3 @@
+pub mod config;
 pub mod init;
 pub mod web;
