+++ src/lib/lib.rs
@@ -1,4 +1,3 @@
 pub mod app;
 pub mod cli;
-pub mod config;
 pub mod util;
